Move table end message to generic component
