+++ src/components/table-infinite-scroll-loader/table-infinite-scroll-loader.styles.ts
