+++ src/components/table-infinite-scroll-loader/table-infinite-scroll-loader.types.ts
@@ -1,5 +1,5 @@
 export type Props = {
-  hasWorkflows: boolean;
+  hasData: boolean;
   error: Error | null;
   fetchNextPage: () => void;
   hasNextPage: boolean;
